requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Airtable
pyairtable>=2.2.0

# Mailchimp
mailchimp-marketing>=3.0.80

# RSS/Feed parsing
feedparser>=6.0.10

# HTML processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data analysis
pandas>=2.0.0

# AI classification
anthropic>=0.18.0

# URL enrichment (optional)
google-generativeai>=0.8.0

# Web scraping for paywalled sites (optional)
playwright>=1.40.0

# Scheduling (optional - for automated runs)
schedule>=1.2.0

# Development
pytest>=7.4.0
